requirements.txt
Cython>=0.22
cmdstanpy==0.9.68
pystan~=2.19.1.1
numpy>=1.15.4
pandas>=1.0.4
matplotlib>=2.0.0
convertdate>=2.1.2
python-dateutil>=2.8.0
performer-pytorch
tqdm>=4.36.1
nystrom-attention
pytorch-lightning